@@ -1,8 +1,8 @@
 #ifndef WAVEFORM_PART_H_INCLUDED
 #define WAVEFORM_PART_H_INCLUDED
 
-#include "exprtk.hpp"
-
+#include <cstdint>
+#include <memory>
 #include <string>
 #include <vector>
 
@@ -12,24 +12,40 @@ namespace fmpire
 class WaveformPart
 {
 public:
-	WaveformPart();
+	enum class Type
+	{
+		SAMPLES,
+		FUNCTION,
+		HARMONIC,
+	};
+
+	static std::shared_ptr<WaveformPart> create(const Type type);
+	static std::shared_ptr<WaveformPart> create(std::string_view& data);
 	virtual ~WaveformPart() noexcept;
 
-	inline bool contains(size_t position) const;
+	inline bool contains(size_t position) const
+	{
+		return (position >= start && position < end);
+	}
+
+	Type get_type() const;
+
 	virtual float sample(size_t position) const = 0;
 
 	virtual std::string encode() const = 0;
-	virtual void decode(const std::string& data) = 0;
+	virtual void decode(std::string_view& data) = 0;
 
 protected:
+	WaveformPart(const Type part_type);
+
+	const Type type;
 	uint32_t waveform_size;
 	uint32_t start;
 	uint32_t end;
 	uint32_t waveform_index;
 
 	std::string encode_generic_part_parameters() const;
-	size_t decode_generic_part_parameters(const std::string& data,
-										  const size_t byte_offset);
+	void decode_generic_part_parameters(std::string_view& data);
 };
 
 class SamplesWaveformPart : public WaveformPart
@@ -41,11 +57,25 @@ public:
 	virtual float sample(size_t position) const override;
 
 	virtual std::string encode() const override;
-	virtual void decode(const std::string& data) override;
+	virtual void decode(std::string_view& data) override;
+
+	inline std::vector<float>& get_samples() { return samples; }
 
 protected:
 	std::vector<float> samples;
 };
+
+} // namespace fmpire
+
+namespace exprtk
+{
+template<typename T> class symbol_table;
+template<typename T> class expression;
+template<typename T> class parser;
+} // namespace exprtk
+
+namespace fmpire
+{
 
 class FunctionWaveformPart : public WaveformPart
 {
@@ -56,14 +86,22 @@ public:
 	virtual float sample(size_t position) const override;
 
 	virtual std::string encode() const override;
-	virtual void decode(const std::string& data) override;
+	virtual void decode(std::string_view& data) override;
+
+	inline void set_function(std::string func)
+	{
+		function = func;
+		update();
+	}
+
+	inline std::string get_function() const { return function; }
 
 protected:
 	std::string function;
 
-	exprtk::symbol_table<float> symbol_table;
-	exprtk::expression<float> expression;
-	exprtk::parser<float> parser;
+	exprtk::symbol_table<float>* symbol_table;
+	exprtk::expression<float>* expression;
+	exprtk::parser<float>* parser;
 	mutable float variable_x;
 	mutable float variable_y;
 
@@ -79,9 +117,9 @@ public:
 	virtual float sample(size_t position) const override;
 
 	virtual std::string encode() const override;
-	virtual void decode(const std::string& data) override;
+	virtual void decode(std::string_view& data) override;
 
-	enum Type
+	enum class HarmonicType
 	{
 		SIN = 0,
 		TRI = 1,
@@ -89,19 +127,21 @@ public:
 		SQR = 3,
 	};
 
-protected:
 	struct Harmonic
 	{
 		float phase;
 		float amplitude;
 	};
 
-	std::vector<Harmonic> harmonics;
-	Type type;
-
-	std::vector<float> samples;
+	inline std::vector<Harmonic>& get_harmonics() { return harmonics; }
 
 	void update();
+
+protected:
+	std::vector<Harmonic> harmonics;
+	HarmonicType type;
+
+	std::vector<float> samples;
 };
 
 } // namespace fmpire
