@@ -2,6 +2,8 @@
 #define WAVEFORM_H_INCLUDED
 
 #include <memory>
+#include <string>
+#include <string_view>
 #include <vector>
 
 namespace fmpire
@@ -17,9 +19,16 @@ public:
 
 	float sample(const size_t position) const;
 
+	std::vector<float> sample_all() const;
+
+	WaveformPart* get_part(size_t position);
+
+	void set_state(std::string_view& state);
+	std::string get_state() const;
+
 private:
-	size_t width;
-	std::vector<WaveformPart> parts;
+	uint32_t width;
+	std::vector<std::shared_ptr<WaveformPart>> parts;
 };
 
 } // namespace fmpire
