@@ -1,42 +1,54 @@
 #include "fmpire_ui.h"
-#include "arc.h"
 
-#include "Color.hpp"
-
-#include "border.h"
-#include "knob.h"
-
-#include <string>
+#include "grid_container.h"
+#include "oscillator_bar.h"
+#include "relative_container.h"
+#include "tooltip.h"
+#include "wavetable_editor.h"
 
 namespace fmpire
 {
 
-FMpireUI::FMpireUI()
+FMpireUI::FMpireUI() :
+	FMpireWindow(this),
+	state_manager(this)
 {
+	setSize(1024, 768);
 	grid = new GridContainer(this);
 	grid->setAbsolutePos(0, 0);
 	grid->setSize(100, 100);
 
-	grid->add_row(1, 0, 3, 0, 150);
+	grid->add_row(1, 0, 3, 0, 100);
 	grid->add_row(3, 0, 0, 0, 0);
 	grid->add_column(1, 0, 0, 0, 0);
 
-	Knob* knob = new Knob(grid);
-	grid->put(knob, 1, 0);
-	knob->set_callback(this);
+	top_bar = new RelativeContainer(grid);
+	grid->put(top_bar, 0, 0);
+
+	tab_selector = new Selector(top_bar);
+	tab_selector->add_option("OSC");
+	tab_selector->add_option("MOD");
+	tab_selector->add_option("FX");
+	tab_selector->add_option("WT");
+	tab_selector->set_callback(this);
+	top_bar->put(tab_selector, 0.25, 0, 0.5, 1);
+
+	oscillator_bar = new OscillatorBar(grid, state_manager);
+	grid->put(oscillator_bar, 1, 0);
+
+	wavetable_editor = new WavetableEditor(grid, state_manager);
+	grid->put(wavetable_editor, 1, 0);
+
+
+	reinit_tooltip();
+	get_tooltip().hide();
+
+	switch_to_tab(0);
+	repaint();
 }
 
 FMpireUI::~FMpireUI() noexcept
 {
-	for (SubWidget* widget : grid->getChildren())
-	{
-		for (SubWidget* widget2 : widget->getChildren())
-		{
-			delete widget2;
-		}
-
-		delete widget;
-	}
 }
 
 void FMpireUI::parameterChanged(uint32_t index, float value)
@@ -46,10 +58,24 @@ void FMpireUI::parameterChanged(uint32_t index, float value)
 void FMpireUI::stateChanged(const char* key, const char* value)
 {
 	d_stdout("state change %s %s", key, value);
+	std::string_view str_view(value);
+	state_manager.state_changed(key, str_view);
+}
+
+void FMpireUI::uiIdle()
+{
+	get_tooltip().idle();
 }
 
 void FMpireUI::onDisplay()
 {
+}
+
+bool FMpireUI::onMotion(const MotionEvent& event)
+{
+	get_tooltip().handle_motion(event);
+	UI::onMotion(event);
+	return false;
 }
 
 void FMpireUI::onResize(const ResizeEvent& ev)
@@ -57,18 +83,44 @@ void FMpireUI::onResize(const ResizeEvent& ev)
 	grid->setSize(ev.size);
 }
 
-void FMpireUI::drag_started(Knob* const knob)
+void FMpireUI::switch_to_tab(const int tab)
 {
+	oscillator_bar->hide();
+	wavetable_editor->hide();
+	switch (tab)
+	{
+	case 0:
+		oscillator_bar->show();
+		break;
+	case 1:
+		break;
+	case 2:
+		break;
+	case 3:
+		wavetable_editor->show();
+		break;
+	default:
+		break;
+	}
+	tab_selector->select(tab);
+	repaint();
 }
 
-void FMpireUI::drag_ended(Knob* const knob)
+void FMpireUI::on_selected(Selector* const selector,
+						   const int index,
+						   const std::string& option)
 {
+	switch_to_tab(index);
 }
-
-void FMpireUI::value_changed(Knob* const knob, float value)
-{
-	setState("vol", std::to_string(value).c_str());
-}
-
 
 } // namespace fmpire
+
+
+START_NAMESPACE_DISTRHO
+
+UI* createUI()
+{
+	return new fmpire::FMpireUI();
+}
+
+END_NAMESPACE_DISTRHO
