@@ -6,7 +6,7 @@ namespace fmpire
 {
 
 Border::Border(Widget* parent) :
-	SubWidget(parent)
+	FMpireWidget(parent)
 {
 }
 
@@ -17,18 +17,33 @@ Border::~Border()
 void Border::onDisplay()
 {
 	const GraphicsContext& context = getGraphicsContext();
-	theme->highlight.setFor(context);
-	float line_width = 2;
-	draw_rounded_box(context, 0, 0, getWidth(), getHeight(), 10, line_width);
+	theme->background.setFor(context);
+	float line_width = theme->line_thin;
+	float radius = theme->corner_radius;
+	draw_rounded_box(context,
+					 0,
+					 0,
+					 getWidth(),
+					 getHeight(),
+					 radius,
+					 line_width);
 }
 
 void Border::onPositionChanged(const PositionChangedEvent& event)
 {
+	if (getChildren().size() == 0)
+	{
+		return;
+	}
 	getChildren().front()->setAbsolutePos(event.pos);
 }
 
 void Border::onResize(const ResizeEvent& event)
 {
+	if (getChildren().size() == 0)
+	{
+		return;
+	}
 	getChildren().front()->setSize(event.size);
 }
 
