@@ -3,6 +3,8 @@
 
 #include "Color.hpp"
 
+#include <string>
+
 namespace fmpire
 {
 
@@ -10,12 +12,20 @@ namespace fmpire
 class Theme
 {
 public:
-	Theme(Color bg, Color fg, Color hl, float thin, float thick) :
+	Theme(Color bg,
+		  Color fg,
+		  Color hl,
+		  float thin,
+		  float thick,
+		  float radius,
+		  std::string font_face) :
 		background(bg),
 		foreground(fg),
 		highlight(hl),
 		line_thin(thin),
-		line_thick(thick)
+		line_thick(thick),
+		corner_radius(radius),
+		font(font_face)
 	{
 	}
 
@@ -24,6 +34,8 @@ public:
 	Color highlight;
 	float line_thin;
 	float line_thick;
+	float corner_radius;
+	std::string font;
 };
 
 class ThemeCollection
